Removed stubs, stacktrace pragma and exporting main
Stubs are not needed since we building with -noStdLib.
This change should have been in the previous commit.

The .stackTrace pragma was not needed in startup.nim
since we have it set in config.nims

Changed from "import" to "include startup" so main,
Reset_Handler and vector table share scope.